将Prometheus时序序列和Metis异常检测结果改成Map匹配并处理key相同的结果、完成异常列表及理由上传接口、完成指标集和指标的对应关系、完成异常检测后通知小组成员的功能、增加日志记录功能、简化数据库表及字段、收集部分指标清单

+++ aiops-service-api/aiops-prometheus-api/src/main/java/com/coocaa/prometheus/feign/ITaskClient.java
@@ -28,5 +28,5 @@ public interface ITaskClient {
     R<Task> createTask(@RequestBody TaskInputVo task, @RequestParam Integer type);
 
     @PostMapping(API_PREFIX + "/remove-task")
-    R<Boolean> removeTask(@RequestParam Integer taskId, @RequestParam Integer type);
+    R<Boolean> removeTask(@RequestParam Long taskId, @RequestParam Integer type);
 }
